misc review comments
